Move models/constants to separate file
